Fix for when nameconflict=overwrite
If the file already existed in the temp directory then the makeunique
would change the file name, so you would not actually overwrite it. Now
if you specify nameconflict=overwrite the cffile upload tag will also
set nameconflict=overwrite to avoid this, otherwise
nameconflict=makeunique WRT to the temp directory.